fix connection issues
